convertServerTimestamp を切り出した
